refactor(welcome): split the welcome window into focused views under Views/Welcome
