--- TablePro/Views/Connection/WelcomeWindowView.swift
@@ -1,745 +1,0 @@
-//
-//  WelcomeWindowView.swift
-//  TablePro
-//
-
-import os
-import SwiftUI
-import TableProImport
-import UniformTypeIdentifiers
-
-struct WelcomeWindowView: View {
-    private enum FocusField {
-        case search
-        case connectionList
-    }
-
-    @State var vm = WelcomeViewModel()
-    @State private var searchFocusTrigger: Int = 0
-    @FocusState private var focus: FocusField?
-
-    var body: some View {
-        ZStack {
-            if vm.showOnboarding {
-                OnboardingContentView {
-                    withMotion(.easeInOut(duration: 0.45)) {
-                        vm.showOnboarding = false
-                    }
-                }
-                .transition(.move(edge: .leading))
-            } else {
-                welcomeContent
-                    .transition(.move(edge: .trailing))
-            }
-        }
-        .ignoresSafeArea()
-        .onAppear {
-            vm.setUp()
-            focus = .connectionList
-        }
-        .alert(
-            vm.connectionsToDelete.count == 1
-                ? String(localized: "Delete Connection")
-                : String(format: String(localized: "Delete %d Connections"), vm.connectionsToDelete.count),
-            isPresented: $vm.showDeleteConfirmation
-        ) {
-            Button(String(localized: "Delete"), role: .destructive) {
-                vm.deleteSelectedConnections()
-            }
-            Button(String(localized: "Cancel"), role: .cancel) {
-                vm.connectionsToDelete = []
-            }
-        } message: {
-            if vm.connectionsToDelete.count == 1, let first = vm.connectionsToDelete.first {
-                if vm.pendingDeleteHasFavorites {
-                    Text("Are you sure you want to delete \"\(first.name)\"? Saved queries linked to this connection will also be deleted.")
-                } else {
-                    Text("Are you sure you want to delete \"\(first.name)\"?")
-                }
-            } else {
-                if vm.pendingDeleteHasFavorites {
-                    Text("Are you sure you want to delete \(vm.connectionsToDelete.count) connections? Saved queries linked to these connections will also be deleted. This cannot be undone.")
-                } else {
-                    Text("Are you sure you want to delete \(vm.connectionsToDelete.count) connections? This cannot be undone.")
-                }
-            }
-        }
-        .alert(
-            String(localized: "Delete Group"),
-            isPresented: $vm.showDeleteGroupConfirmation
-        ) {
-            Button(String(localized: "Delete"), role: .destructive) {
-                vm.confirmDeleteGroup()
-            }
-            Button(String(localized: "Cancel"), role: .cancel) {
-                vm.groupToDelete = nil
-            }
-        } message: {
-            if let group = vm.groupToDelete {
-                Text("Are you sure you want to delete the group \"\(group.name)\"? Connections in this group will be moved to the top level.")
-            }
-        }
-        .sheet(item: $vm.activeSheet, onDismiss: {
-            if let count = vm.pendingImportResultCount {
-                vm.importResultCount = count
-                vm.pendingImportResultCount = nil
-            }
-            focus = .connectionList
-            WindowOpener.shared.openStagedConnectionForm()
-        }) { sheet in
-            switch sheet {
-            case .newGroup(let parentId):
-                CreateGroupSheet(parentId: parentId) { name, color, pid in
-                    try vm.createGroup(name: name, color: color, parentId: pid)
-                }
-            case .activation:
-                LicenseActivationSheet()
-            case .importFile(let url):
-                ConnectionImportSheet(fileURL: url) { count in
-                    vm.pendingImportResultCount = count
-                    vm.activeSheet = nil
-                }
-            case .exportConnections(let conns):
-                ConnectionExportOptionsSheet(connections: conns)
-            case .importFromApp:
-                ImportFromAppSheet { count in
-                    vm.pendingImportResultCount = count
-                    vm.activeSheet = nil
-                }
-            case .projectFolderScan(let url):
-                ProjectFolderScanSheet(
-                    rootURL: url,
-                    onSelect: { parsed in
-                        vm.activeSheet = nil
-                        WindowOpener.shared.stageConnectionFormDraft(parsedURL: parsed)
-                    },
-                    onChooseAnotherFolder: {
-                        vm.activeSheet = nil
-                        vm.openProjectFolder()
-                    }
-                )
-            case .deeplinkImport(let exportable):
-                DeeplinkImportSheet(connection: exportable) {
-                    vm.loadConnections()
-                }
-            }
-        }
-        .modifier(ConnectionCreationOverlays(vm: vm))
-        .pluginInstallPromptForType(type: $vm.pendingInstallType) { type in
-            vm.completePendingInstall(for: type)
-        }
-        .pluginInstallPrompt(connection: $vm.pluginInstallConnection) { connection in
-            vm.connectAfterInstall(connection)
-        }
-        .sheet(item: $vm.pluginDiagnostic) { item in
-            PluginDiagnosticSheet(item: item) {
-                vm.pluginDiagnostic = nil
-            }
-        }
-        .alert(String(localized: "Rename Group"), isPresented: $vm.showRenameGroupAlert) {
-            TextField(String(localized: "Group name"), text: $vm.renameGroupName)
-            Button(String(localized: "Rename")) { vm.confirmRenameGroup() }
-            Button(String(localized: "Cancel"), role: .cancel) { vm.renameGroupTarget = nil }
-        } message: {
-            Text("Enter a new name for the group.")
-        }
-        .alert(
-            String(localized: "Group Not Updated"),
-            isPresented: Binding(
-                get: { vm.groupErrorMessage != nil },
-                set: { if !$0 { vm.groupErrorMessage = nil } }
-            )
-        ) {
-            Button(String(localized: "OK")) { vm.groupErrorMessage = nil }
-        } message: {
-            if let message = vm.groupErrorMessage {
-                Text(message)
-            }
-        }
-        .alert(
-            String(localized: "Connection Failed"),
-            isPresented: $vm.showConnectionError
-        ) {
-            Button(String(localized: "OK"), role: .cancel) {
-                vm.connectionError = nil
-            }
-        } message: {
-            if let error = vm.connectionError {
-                Text(error)
-            }
-        }
-        .fileImporter(
-            isPresented: $vm.showImportFilePanel,
-            allowedContentTypes: [.tableproConnectionShare],
-            allowsMultipleSelection: false
-        ) { result in
-            if case .success(let urls) = result, let url = urls.first {
-                vm.activeSheet = .importFile(url)
-            }
-        }
-        .alert(
-            (vm.importResultCount ?? 0) > 0
-                ? String(localized: "Import Complete")
-                : String(localized: "No Connections Imported"),
-            isPresented: Binding(
-                get: { vm.importResultCount != nil },
-                set: { if !$0 { vm.importResultCount = nil } }
-            )
-        ) {
-            Button(String(localized: "OK")) { vm.importResultCount = nil }
-        } message: {
-            if let count = vm.importResultCount, count > 0 {
-                Text(count == 1
-                    ? String(localized: "1 connection was imported.")
-                    : String(format: String(localized: "%d connections were imported."), count))
-            } else {
-                Text(String(localized: "All selected connections were skipped."))
-            }
-        }
-    }
-
-    // MARK: - Layout
-
-    private var welcomeContent: some View {
-        HStack(spacing: 0) {
-            WelcomeActionsPanel(
-                onActivateLicense: { vm.activeSheet = .activation },
-                onCreateConnection: { WindowOpener.shared.openConnectionForm() },
-                onImportFromURL: { vm.urlImportPresented = true },
-                onImportFromApp: { vm.importConnectionsFromApp() },
-                onOpenProjectFolder: { vm.openProjectFolder() },
-                onImportConnectionsFile: { vm.importConnectionsFromFile() }
-            )
-            .frame(width: 240)
-            .themeMaterial(.sidebar, .regularMaterial)
-
-            Divider()
-
-            connectionsPanel
-        }
-        .transition(.opacity)
-    }
-
-    // MARK: - Connections panel
-
-    private var connectionsPanel: some View {
-        VStack(spacing: 0) {
-            connectionsHeader
-            Divider()
-            if !vm.availableTags.isEmpty {
-                TagFilterBar(tagFilter: $vm.tagFilter, availableTags: vm.availableTags)
-                Divider()
-            }
-            ZStack {
-                if vm.treeItems.isEmpty && vm.linkedConnections.isEmpty && vm.teamLibraryConnections.isEmpty && vm.favoriteConnections.isEmpty {
-                    emptyState
-                } else {
-                    connectionList
-                }
-            }
-        }
-        .frame(maxWidth: .infinity, maxHeight: .infinity)
-        .background(Color(nsColor: .controlBackgroundColor))
-        .contentShape(Rectangle())
-        .contextMenu { newConnectionContextMenu }
-        .onReceive(NotificationCenter.default.publisher(for: .welcomeWindowFindRequested)) { _ in
-            searchFocusTrigger += 1
-        }
-    }
-
-    private var newConnectionHelp: String {
-        let binding = AppSettingsManager.shared.keyboard.shortcut(for: .newConnection)
-        guard let displayString = binding?.displayString, !displayString.isEmpty else {
-            return String(localized: "New Connection")
-        }
-        return String(format: String(localized: "New Connection (%@)"), displayString)
-    }
-
-    private var connectionsHeader: some View {
-        HStack(spacing: 8) {
-            Button {
-                WindowOpener.shared.openConnectionForm()
-            } label: {
-                Image(systemName: "plus")
-                    .frame(width: 14, height: 14)
-            }
-            .buttonStyle(.bordered)
-            .controlSize(.large)
-            .help(newConnectionHelp)
-            .accessibilityLabel(String(localized: "New Connection"))
-
-            Button {
-                vm.pendingMoveToNewGroup = []
-                vm.activeSheet = .newGroup(parentId: nil)
-            } label: {
-                Image(systemName: "folder.badge.plus")
-                    .frame(width: 14, height: 14)
-            }
-            .buttonStyle(.bordered)
-            .controlSize(.large)
-            .help(String(localized: "New Group"))
-            .accessibilityLabel(String(localized: "New Group"))
-
-            Spacer()
-
-            NativeSearchField(
-                text: $vm.searchText,
-                placeholder: String(localized: "Search for connection…"),
-                controlSize: .regular,
-                onMoveDown: { focus = .connectionList },
-                onSubmit: { focus = .connectionList },
-                focusTrigger: searchFocusTrigger,
-                maxWidth: 240
-            )
-            .focused($focus, equals: .search)
-            .layoutPriority(0)
-        }
-        .padding(.horizontal, 12)
-        .padding(.vertical, 10)
-    }
-
-    // MARK: - Connection List
-
-    private var connectionList: some View {
-        ScrollViewReader { proxy in
-            List(selection: $vm.selectedConnectionIds) {
-                let showsFavoritesSection = vm.searchText.isEmpty && !vm.favoriteConnections.isEmpty
-                let showsLinkedSection = !vm.linkedConnections.isEmpty
-                    && LicenseManager.shared.isFeatureAvailable(.linkedFolders)
-                let showsTeamLibrarySection = !vm.teamLibraryConnections.isEmpty
-                    && LicenseManager.shared.isFeatureAvailable(.teamLibrary)
-                let treeHasGroups = vm.treeItems.contains { item in
-                    if case .group = item { return true }
-                    return false
-                }
-                let treeNeedsHeader = showsFavoritesSection && !treeHasGroups && !vm.treeItems.isEmpty
-
-                if showsFavoritesSection {
-                    Section {
-                        ForEach(vm.favoriteConnections) { conn in
-                            connectionRow(for: conn)
-                        }
-                    } header: {
-                        sourceListSectionHeader(String(localized: "Favorites"))
-                    }
-                }
-
-                if treeNeedsHeader {
-                    Section {
-                        TreeRowsView(items: vm.treeItems, parentGroupId: nil, vm: vm) { conn in
-                            connectionRow(for: conn)
-                        }
-                    } header: {
-                        sourceListSectionHeader(String(localized: "Connections"))
-                    }
-                } else {
-                    TreeRowsView(items: vm.treeItems, parentGroupId: nil, vm: vm) { conn in
-                        connectionRow(for: conn)
-                    }
-                }
-
-                if showsLinkedSection {
-                    Section {
-                        ForEach(vm.linkedConnections) { linked in
-                            linkedConnectionRow(for: linked)
-                        }
-                    } header: {
-                        sourceListSectionHeader(String(localized: "Linked"))
-                    }
-                }
-
-                if showsTeamLibrarySection {
-                    Section {
-                        ForEach(vm.teamLibraryConnections) { linked in
-                            teamLibraryConnectionRow(for: linked)
-                        }
-                    } header: {
-                        sourceListSectionHeader(String(localized: "Team Library"))
-                    }
-                }
-            }
-            .listStyle(.inset)
-            .listRowSeparator(.hidden)
-            .listSectionSeparator(.hidden)
-            .scrollContentBackground(.hidden)
-            .focused($focus, equals: .connectionList)
-            .contextMenu(forSelectionType: UUID.self) { ids in
-                contextMenuContent(for: ids)
-            } primaryAction: { ids in
-                primaryAction(for: ids)
-            }
-            .onKeyPress(characters: .init(charactersIn: "\u{7F}\u{08}"), phases: .down) { keyPress in
-                guard keyPress.modifiers.contains(.command) else { return .ignored }
-                let toDelete = vm.selectedConnections
-                guard !toDelete.isEmpty else { return .ignored }
-                vm.requestDeleteConnections(toDelete)
-                return .handled
-            }
-            .onKeyPress(characters: .init(charactersIn: "a"), phases: .down) { keyPress in
-                guard keyPress.modifiers.contains(.command) else { return .ignored }
-                vm.selectedConnectionIds = Set(vm.flatVisibleConnections.map(\.id))
-                return .handled
-            }
-            .onKeyPress(.escape) {
-                if !vm.selectedConnectionIds.isEmpty {
-                    vm.selectedConnectionIds = []
-                }
-                return .handled
-            }
-            .onKeyPress(characters: .init(charactersIn: "jn"), phases: [.down, .repeat]) { keyPress in
-                guard keyPress.modifiers.contains(.control) else { return .ignored }
-                vm.moveToNextConnection()
-                scrollToSelection(proxy)
-                return .handled
-            }
-            .onKeyPress(characters: .init(charactersIn: "kp"), phases: [.down, .repeat]) { keyPress in
-                guard keyPress.modifiers.contains(.control) else { return .ignored }
-                vm.moveToPreviousConnection()
-                scrollToSelection(proxy)
-                return .handled
-            }
-            .onKeyPress(characters: .init(charactersIn: "h"), phases: .down) { keyPress in
-                guard keyPress.modifiers.contains(.control) else { return .ignored }
-                vm.collapseSelectedGroup()
-                return .handled
-            }
-            .onKeyPress(characters: .init(charactersIn: "l"), phases: .down) { keyPress in
-                guard keyPress.modifiers.contains(.control) else { return .ignored }
-                vm.expandSelectedGroup()
-                return .handled
-            }
-        }
-    }
-
-    // MARK: - Rows
-
-    func connectionRow(for connection: DatabaseConnection) -> some View {
-        let sshProfile = connection.sshProfileId.flatMap { SSHProfileStorage.shared.profile(for: $0) }
-        return WelcomeConnectionRow(
-            connection: connection,
-            sshProfile: sshProfile,
-            isSelected: vm.selectedConnectionIds.contains(connection.id),
-            onToggleFavorite: { vm.toggleFavorite([connection]) }
-        )
-        .tag(connection.id)
-        .listRowSeparator(.hidden)
-        .swipeActions(edge: .trailing, allowsFullSwipe: false) {
-            editConnectionButton(for: connection)
-            deleteConnectionButton(for: connection)
-        }
-    }
-
-    private func sourceListSectionHeader(_ title: String) -> some View {
-        Text(title)
-            .font(.subheadline)
-            .fontWeight(.semibold)
-            .foregroundStyle(.secondary)
-            .padding(.top, 6)
-    }
-
-    private func linkedConnectionRow(for linked: LinkedConnection) -> some View {
-        HStack(spacing: 12) {
-            ZStack(alignment: .bottomTrailing) {
-                DatabaseType(rawValue: linked.connection.type).iconImage
-                    .frame(width: 28, height: 28)
-                Image(systemName: "folder.fill")
-                    .font(.caption2)
-                    .foregroundStyle(.secondary)
-                    .offset(x: 2, y: 2)
-            }
-            VStack(alignment: .leading, spacing: 2) {
-                Text(linked.connection.name)
-                    .lineLimit(1)
-                Text(verbatim: linked.connection.displaySubtitle)
-                    .font(.caption)
-                    .foregroundStyle(.secondary)
-                    .lineLimit(1)
-            }
-        }
-        .tag(linked.id)
-        .contentShape(Rectangle())
-        .listRowSeparator(.hidden)
-    }
-
-    private func teamLibraryConnectionRow(for linked: LinkedConnection) -> some View {
-        HStack(spacing: 12) {
-            ZStack(alignment: .bottomTrailing) {
-                DatabaseType(rawValue: linked.connection.type).iconImage
-                    .frame(width: 28, height: 28)
-                Image(systemName: "person.2.fill")
-                    .font(.caption2)
-                    .foregroundStyle(.secondary)
-                    .offset(x: 2, y: 2)
-            }
-            VStack(alignment: .leading, spacing: 2) {
-                Text(linked.connection.name)
-                    .lineLimit(1)
-                Text(verbatim: linked.connection.displaySubtitle)
-                    .font(.caption)
-                    .foregroundStyle(.secondary)
-                    .lineLimit(1)
-            }
-        }
-        .tag(linked.id)
-        .contentShape(Rectangle())
-        .listRowSeparator(.hidden)
-    }
-
-    func primaryAction(for ids: Set<UUID>) {
-        guard !ids.isEmpty else { return }
-        for connection in vm.connections where ids.contains(connection.id) {
-            vm.connectToDatabase(connection)
-        }
-        for linked in vm.linkedConnections where ids.contains(linked.id) {
-            vm.connectToLinkedConnection(linked)
-        }
-        for linked in vm.teamLibraryConnections where ids.contains(linked.id) {
-            vm.connectToLinkedConnection(linked)
-        }
-    }
-
-    // MARK: - Empty State
-
-    @ViewBuilder
-    private var emptyState: some View {
-        if vm.searchText.isEmpty {
-            EmptyStateView(
-                icon: "cylinder.split.1x2",
-                title: String(localized: "No Connections"),
-                description: String(localized: "Try the sample database, or click + above to add your own."),
-                actionTitle: String(localized: "Try Sample Database"),
-                action: { vm.openSampleDatabase() }
-            )
-            .frame(maxWidth: .infinity, maxHeight: .infinity)
-        } else {
-            EmptyStateView(
-                icon: "magnifyingglass",
-                title: String(localized: "No Matching Connections"),
-                description: String(localized: "Try a different search term.")
-            )
-            .frame(maxWidth: .infinity, maxHeight: .infinity)
-        }
-    }
-
-    // MARK: - Helpers
-
-    private func scrollToSelection(_ proxy: ScrollViewProxy) {
-        if let id = vm.selectedConnectionIds.first {
-            proxy.scrollTo(id, anchor: .center)
-        }
-    }
-}
-
-// MARK: - Tree Rendering
-
-private struct TreeRowsView<ConnectionContent: View>: View {
-    let items: [ConnectionGroupTreeNode]
-    let parentGroupId: UUID?
-    var vm: WelcomeViewModel
-    let connectionRowBuilder: (DatabaseConnection) -> ConnectionContent
-
-    private var hasGroups: Bool {
-        items.contains { node in
-            if case .group = node { return true }
-            return false
-        }
-    }
-
-    var body: some View {
-        let allConnections = !hasGroups
-        ForEach(items) { item in
-            switch item {
-            case .connection(let conn):
-                connectionRowBuilder(conn)
-            case .group(let group, let children):
-                DisclosureGroup(isExpanded: expandedBinding(for: group.id)) {
-                    TreeRowsView(
-                        items: children,
-                        parentGroupId: group.id,
-                        vm: vm,
-                        connectionRowBuilder: connectionRowBuilder
-                    )
-                } label: {
-                    groupLabel(for: group)
-                }
-                .listRowSeparator(.hidden)
-            }
-        }
-        .onMove(perform: allConnections ? { from, to in
-            guard vm.searchText.isEmpty else { return }
-            vm.moveConnections(
-                renderedIds: items.compactMap { item in
-                    guard case .connection(let conn) = item else { return nil }
-                    return conn.id
-                },
-                from: from,
-                to: to,
-                inGroup: parentGroupId
-            )
-        } : nil)
-    }
-
-    private func expandedBinding(for groupId: UUID) -> Binding<Bool> {
-        Binding(
-            get: { vm.expandedGroupIds.contains(groupId) },
-            set: { expanded in
-                if expanded {
-                    vm.expandedGroupIds.insert(groupId)
-                } else {
-                    vm.expandedGroupIds.remove(groupId)
-                }
-            }
-        )
-    }
-
-    private func groupLabel(for group: ConnectionGroup) -> some View {
-        HStack(spacing: 6) {
-            if !group.color.isDefault {
-                Circle()
-                    .fill(group.color.color)
-                    .frame(width: 8, height: 8)
-            }
-
-            Text(group.name)
-                .font(.subheadline.weight(.semibold))
-                .foregroundStyle(.secondary)
-
-            Text("\(vm.connectionCountByGroup[group.id] ?? 0)")
-                .font(.caption2)
-                .foregroundStyle(.tertiary)
-
-            Spacer()
-        }
-        .contentShape(Rectangle())
-        .contextMenu {
-            groupContextMenu(for: group)
-        }
-    }
-
-    @ViewBuilder
-    private func groupContextMenu(for group: ConnectionGroup) -> some View {
-        Button {
-            vm.beginRenameGroup(group)
-        } label: {
-            Label(String(localized: "Rename"), systemImage: "pencil")
-        }
-
-        let currentGroupDepth = vm.depthByGroup[group.id] ?? 0
-        Button {
-            vm.createSubgroup(under: group.id)
-        } label: {
-            Label(String(localized: "New Subgroup"), systemImage: "folder.badge.plus")
-        }
-        .disabled(currentGroupDepth >= 3)
-
-        Menu(String(localized: "Change Color")) {
-            ForEach(ConnectionColor.allCases) { color in
-                Button {
-                    vm.updateGroupColor(group, color: color)
-                } label: {
-                    HStack {
-                        if color != .none {
-                            Image(systemName: "circle.fill")
-                                .foregroundStyle(color.color)
-                        }
-                        Text(color.displayName)
-                        if group.color == color {
-                            Spacer()
-                            Image(systemName: "checkmark")
-                        }
-                    }
-                }
-            }
-        }
-
-        if vm.groups.count > 1 {
-            Menu(String(localized: "Move Group to…")) {
-                Button {
-                    vm.moveGroup(group, toParent: nil)
-                } label: {
-                    HStack {
-                        Text("Top Level")
-                        if group.parentId == nil {
-                            Spacer()
-                            Image(systemName: "checkmark")
-                        }
-                    }
-                }
-                .disabled(group.parentId == nil)
-
-                Divider()
-
-                ForEach(vm.groups.filter({ $0.id != group.id })) { targetGroup in
-                    let canPlace = canPlaceGroup(group.id, under: targetGroup.id, groups: vm.groups)
-
-                    Button {
-                        vm.moveGroup(group, toParent: targetGroup.id)
-                    } label: {
-                        HStack {
-                            if !targetGroup.color.isDefault {
-                                Image(systemName: "circle.fill")
-                                    .foregroundStyle(targetGroup.color.color)
-                            }
-                            Text(targetGroup.name)
-                            if group.parentId == targetGroup.id {
-                                Spacer()
-                                Image(systemName: "checkmark")
-                            }
-                        }
-                    }
-                    .disabled(!canPlace || group.parentId == targetGroup.id)
-                }
-            }
-        }
-
-        Divider()
-
-        Button(role: .destructive) {
-            vm.requestDeleteGroup(group)
-        } label: {
-            Label(String(localized: "Delete Group"), systemImage: "trash")
-        }
-    }
-}
-
-// MARK: - Connection Creation Overlays
-
-private struct ConnectionCreationOverlays: ViewModifier {
-    @Bindable var vm: WelcomeViewModel
-
-    func body(content: Content) -> some View {
-        content
-            .sheet(item: $vm.databaseTypeChooser, onDismiss: {
-                WindowOpener.shared.openStagedConnectionForm()
-            }) { payload in
-                DatabaseTypeChooserSheet(
-                    initialType: payload.initialType,
-                    onSelected: { type in
-                        vm.selectDatabaseType(type, for: payload)
-                    },
-                    onImportFromURL: { vm.presentURLImport() },
-                    onCancel: { vm.databaseTypeChooser = nil }
-                )
-            }
-            .sheet(isPresented: $vm.urlImportPresented, onDismiss: {
-                WindowOpener.shared.openStagedConnectionForm()
-            }) {
-                ImportFromURLSheet(
-                    onImported: { parsed in
-                        vm.urlImportPresented = false
-                        WindowOpener.shared.stageConnectionFormDraft(parsedURL: parsed)
-                    },
-                    onCancel: {
-                        vm.urlImportPresented = false
-                    }
-                )
-            }
-    }
-}
-
-// MARK: - Preview
-
-#Preview("Welcome Window") {
-    WelcomeWindowView()
-        .frame(width: 700, height: 450)
-}
