+++ TablePro/Views/Welcome/WelcomeActionsPanel.swift
